+++ app/Model/User.php
@@ -106,6 +106,7 @@ class User extends AppModel {
 	public $hasMany = [
 		'GroupUser',
 		'Secret',
+		'UserResourcePermission',
 		// Custom join with ControllerLog to retrieve the last logged in date.
 		// The results of this will be processed in the afterFind
 		// and integrated directly in the user object with column name last_logged_in.
@@ -609,6 +610,12 @@ class User extends AppModel {
 								'Role.name' => [Role::USER, Role::ADMIN],
 							]
 						];
+						// if user is admin, is-active filter is enabled
+						if ($role == Role::ADMIN) {
+							if (isset($data['filter']['is-active'])) {
+								$conditions['conditions']['User.active'] = $data['filter']['is-active'] ? 1 : 0;
+							}
+						}
 						// if user is simple user, we do not allow him to see non active users.
 						if ($role == Role::USER) {
 							$conditions['conditions']['User.active'] = 1;